+++ R/test_correlation.R
@@ -6,6 +6,15 @@
 #' @param alpha Significance level.
 #' @param plot Logical; include a ggplot object.
 #' @param na.rm Logical; remove missing values.
+#' @references
+#' Pearson, K. (1895). Notes on regression and inheritance in the case of two
+#' parents. \emph{Proceedings of the Royal Society of London}, 58, 240-242.
+#'
+#' Spearman, C. (1904). The proof and measurement of association between two
+#' things. \emph{The American Journal of Psychology}, 15(1), 72-101.
+#'
+#' Kendall, M. G. (1938). A new measure of rank correlation. \emph{Biometrika},
+#' 30(1/2), 81-93.
 #' @export
 test_correlation <- function(formula, data, y = NULL, method = c("auto", "pearson", "spearman", "kendall"), alpha = 0.05, plot = TRUE, na.rm = TRUE) {
   method <- match.arg(method)